backend/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
django-cors-headers==3.13.0
pillow==9.3.0
drf-extra-fields==3.4.1
drf-writable-nested==0.7.0